IntegratedWienerProcesses

Documentation for IntegratedWienerProcesses.

IntegratedWienerProcesses.IWP

IntegratedWienerProcesses._iwp_ssparams_reverse_1d

IntegratedWienerProcesses._rgram_cholf_reverse_1d

IntegratedWienerProcesses._transition_matrix_1d

IntegratedWienerProcesses.dim

IntegratedWienerProcesses.ndiff

IntegratedWienerProcesses.ssparams_reverse

IntegratedWienerProcesses.state2diff_matrix_reverse

IntegratedWienerProcesses.transition_parameters_cholf_reverse

IntegratedWienerProcesses.IWP — Method

IWP{U}(M::IWP{T,N})

Computes an IWP of element type U from the IWP M of element type T

IntegratedWienerProcesses._iwp_ssparams_reverse_1d — Method

_iwp_ssparams_reverse_1d(::Type{T}, ndiff::Integer)

Computes the matrices A, B, and C associated with the state-space realization of a one dimensiona IWP, where the state vector x is ordered from highest Taylor coefficient to the lowest.

IntegratedWienerProcesses._rgram_cholf_reverse_1d — Method

_rgram_cholf_reverse_1d(ndiff::Integer, dt::T)

Computes the lower triangular Cholesky factor reachability Grammian of ndiff-fold integrator over the interval [0, dt]. This is the same as the IWP process noise for a transtion over [t, t+dt].

IntegratedWienerProcesses._transition_matrix_1d — Method

_transition_matrix_1d(ndiff::Integer, dt::T)

Computes the transition matrix A of an ndiff-fold integrator over the interval [0, dt].

IntegratedWienerProcesses.dim — Method

dim(M::IWP)

Computes the number of dimensions of M.

IntegratedWienerProcesses.ndiff — Method

ndiff(M::IWP)

Computes the number of times M is differentiable.

IntegratedWienerProcesses.ssparams_reverse — Method

ssparams_reverse(M::IWP{T})

Computes the matrices A, B, and C associated with the state-space realization of M

dx = Axdt + Bdw y = Cx

where the state vector x is ordered from highest Taylor coefficient to the lowest.

IntegratedWienerProcesses.state2diff_matrix_reverse — Method

state2diff_matrix_reverse(M::IWP{T}, m::Integer)

Computes the matrix Cm such that

D^m y = Cm*x,

where x is the state vector of the reverse ordered state-space realization.

IntegratedWienerProcesses.transition_parameters_cholf_reverse — Method

transition_parameters_cholf_reverse(M::IWP{T}, dt::T)

Computes the transition parameters A, L of the IWP M for the interval dt. Here L is the lower triangular Cholesky factor of the process noise,