IntegratedWienerProcesses

Documentation for IntegratedWienerProcesses.

IntegratedWienerProcesses.ReverseTaylor

IntegratedWienerProcesses.dim

IntegratedWienerProcesses.ndiff

IntegratedWienerProcesses.ssparams

IntegratedWienerProcesses.state2diff_matrix

IntegratedWienerProcesses.transition_cov

IntegratedWienerProcesses.transition_cov_cholf

IntegratedWienerProcesses.transition_matrix

IntegratedWienerProcesses.ReverseTaylor — Type

ReverseTaylor

Type for representing state-space realizations in terms of Taylor coefficients in reverse order. That is the state-vector is of the form:

x = [D^n y, D^(n-1)y, ..., Dy, y]

IntegratedWienerProcesses.dim — Method

dim(M::IWP)

Computes the number of dimensions of M.

IntegratedWienerProcesses.ndiff — Method

ndiff(M::IWP)

Computes the number of times M is differentiable.

IntegratedWienerProcesses.ssparams — Method

ssparams(model::IWP{T}, method::AbstractRealizationMethod) where {T}

Computes the state-space matrices A, B, C for model model using realization method method.

IntegratedWienerProcesses.state2diff_matrix — Method

state2diff_matrix(model::IWP{T}, m::Integer, method::AbstractRealizationMethod) where {T}

Computes the matrix Em such that Dy = Em * x, where x is the state vector associated with realization method method.

IntegratedWienerProcesses.transition_cov — Method

transition_cov(model::IWP{T}, dt::T, method::AbstractRealizationMethod) where {T}

Computes the transition covariance of the model model, with step-size dt, using realization method method.

IntegratedWienerProcesses.transition_cov_cholf — Method

transition_cov_cholf(model::IWP{T}, dt::T, method::AbstractRealizationMethod) where {T}

Computes a left cholesky factor of the transition covariance of the model model, with step-size dt, using realization method method.

IntegratedWienerProcesses.transition_matrix — Method

transition_matrix(model::IWP{T}, dt::Real, method::AbstractRealizationMethod) where {T}

Computes the transition matrix of model model, using step-size dt, and realization method method.